FIX: Updates mni_atlas_reader to atlasreader
The name change impacted a lot of stuff! Thankfully find-and-replace is
a magical tool. I _think_ I checked everything to ensure that I wasn't
renaming things that shouldn't be renamed, but a once over might be
good. And if there are a few things that need to be fixed in later PRs,
so be it!