[FIX] rounding issue
Dear all,

I found a bug for an edge condition in atlasreader. I think this is due to a rounding error (or rather no rounding but only changing float to integers without rounding at all). The two tables below show the peak table with the current implementation of atlasreader, where the Amygdala peak has a 1-p value of 0.  After implementing rounding of the coordinates the Amygdala gets the correct value of 1-p = 0.9996.

I hope this help,
Anna

Before the fix:
	cluster_id	peak_x	peak_y	peak_z	peak_value	volume_mm	harvard_oxford
0	1	21.5	-12.5	-16.5	0.9998	6744	81% Right_Hippocampus
1	1	27.5	-34.5	-0.5	0.9990	6744	62% Right_Hippocampus
2	2	-24.5	-14.5	-16.5	0.9998	5816	92% Left_Hippocampus
3	2	-22.5	-6.5	-28.5	0.9812	5816	36% Left_Hippocampus; 30% Left_Parahippocampal_Gyrus_anterior_division
4	2	-32.5	-2.5	-24.5	0.0000	5816	25% Left_Amygdala


After rounding floats before converting them to integers:
0 	1 	23.5 	-10.5 	-16.5 	0.9998 	6744 	38% Right_Hippocampus; 25% Right_Amygdala
1 	1 	27.5 	-34.5 	-0.5 	0.9990 	6744 	54% Right_Hippocampus
2 	2 	-24.5 	-14.5 	-14.5 	0.9998 	5816 	42% Left_Hippocampus
3 	2 	-30.5 	-0.5 	-24.5 	0.9996 	5816 	58% Left_Amygdala
4 	2 	-22.5 	-6.5 	-28.5 	0.9812 	5816 	47% Left_Hippocampus